Refactoring a little more so the AMI resolver and provider are separated better
