+++ pkg/cloudprovider/aws/amifamily/ami.go
@@ -29,6 +29,7 @@ import (
 	"github.com/aws/karpenter/pkg/apis/awsnodetemplate/v1alpha1"
 	"github.com/aws/karpenter/pkg/apis/provisioning/v1alpha5"
 	"github.com/aws/karpenter/pkg/cloudprovider"
+	awsv1alpha1 "github.com/aws/karpenter/pkg/cloudprovider/aws/apis/v1alpha1"
 	"github.com/aws/karpenter/pkg/scheduling"
 )
 
@@ -39,10 +40,34 @@ type AMIProvider struct {
 }
 
 // Get returns a set of AMIIDs and corresponding instance types. AMI may vary due to architecture, accelerator, etc
-func (p *AMIProvider) Get(ctx context.Context, instanceType cloudprovider.InstanceType, ssmQuery string, amiRequirements map[string]scheduling.Requirements) (string, error) {
-	if len(amiRequirements) > 0 {
-		return amiOverride(instanceType, amiRequirements), nil
+// If AMI overrides are specified in the AWSNodeTemplate, then only those AMIs will be chosen.
+func (p *AMIProvider) Get(ctx context.Context, provider *awsv1alpha1.AWS, nodeRequest *cloudprovider.NodeRequest, options *Options, amiFamily AMIFamily) (map[string][]cloudprovider.InstanceType, error) {
+	amiIDs := map[string][]cloudprovider.InstanceType{}
+	amiRequirements, err := p.GetAMIRequirements(ctx, nodeRequest.Template.ProviderRef)
+	if err != nil {
+		return nil, err
 	}
+	for _, instanceType := range nodeRequest.InstanceTypeOptions {
+		var amiID string
+		if len(amiRequirements) > 0 {
+			amiID = getAMIOverride(instanceType, amiRequirements)
+		} else {
+			amiID, err = p.GetDefaultAMIFromSSM(ctx, instanceType, amiFamily.SSMAlias(options.KubernetesVersion, instanceType))
+			if err != nil {
+				return nil, err
+			}
+		}
+		if amiID != "" {
+			amiIDs[amiID] = append(amiIDs[amiID], instanceType)
+		}
+	}
+	if len(amiIDs) == 0 {
+		return nil, fmt.Errorf("no instance types satisfy requirements of amis %v,", Keys(amiRequirements))
+	}
+	return amiIDs, nil
+}
+
+func (p *AMIProvider) GetDefaultAMIFromSSM(ctx context.Context, instanceType cloudprovider.InstanceType, ssmQuery string) (string, error) {
 	if id, ok := p.cache.Get(ssmQuery); ok {
 		return id.(string), nil
 	}
@@ -70,11 +95,20 @@ func (p *AMIProvider) GetAMIRequirements(ctx context.Context, providerRef *v1alp
 	return amiRequirements, nil
 }
 
-func amiOverride(instanceType cloudprovider.InstanceType, amiRequirements map[string]scheduling.Requirements) string {
+func getAMIOverride(instanceType cloudprovider.InstanceType, amiRequirements map[string]scheduling.Requirements) string {
 	for amiID, requirements := range amiRequirements {
 		if err := instanceType.Requirements().Compatible(requirements); err == nil {
 			return amiID
 		}
 	}
 	return ""
+}
+
+// Keys returns a slice of all the keys in a map
+func Keys(m map[string]scheduling.Requirements) []string {
+	keys := make([]string, 0, len(m))
+	for key := range m {
+		keys = append(keys, key)
+	}
+	return keys
 }
